CLN : literalinclude default_traits_view and multiple_views
	modified:   docs/source/traitsui_user_manual/advanced_view.rst
	new file:   docs/source/traitsui_user_manual/examples/default_traits_view.py
	new file:   docs/source/traitsui_user_manual/examples/multiple_views.py
	deleted:    examples/tutorials/doc_examples/examples/default_traits_view.py
	deleted:    examples/tutorials/doc_examples/examples/multiple_views.py
