+++ docs/source/traitsui_user_manual/advanced_view.rst
@@ -63,34 +63,8 @@ variation on Example 3:
 
 .. rubric:: Example 5: Using configure_traits() with a default View object
 
-::
-
-    # default_traits_view.py -- Sample code to demonstrate the use of
-    # 'traits_view'
-    from traits.api import HasTraits, Str, Int
-    from traitsui.api import View, Item, Group
-    import traitsui
-
-    class SimpleEmployee2(HasTraits):
-        first_name = Str()
-        last_name = Str()
-        department = Str()
-
-        employee_number = Str()
-        salary = Int()
-
-        traits_view = View(
-            Group(
-                Item(name='first_name'),
-                Item(name='last_name'),
-                Item(name='department'),
-                label='Personnel profile',
-                show_border=True,
-            ),
-        )
-
-    sam = SimpleEmployee2()
-    sam.configure_traits()
+.. literalinclude:: examples/default_traits_view.py
+   :start-at: default_traits_view.py
 
 In this example, configure_traits() no longer requires a *view* keyword
 argument; the **traits_view** attribute is used by default, resulting in the
@@ -173,47 +147,8 @@ this by simply adding a second View attribute:
 
 .. rubric:: Example 6: Defining multiple View objects in a HasTraits class
 
-::
-
-    # multiple_views.py -- Sample code to demonstrate the use of
-    # multiple views
-    from traits.api import HasTraits, Str, Int
-    from traitsui.api import View, Item, Group
-    import traitsui
-
-    class SimpleEmployee3(HasTraits):
-        first_name = Str()
-        last_name = Str()
-        department = Str()
-
-        employee_number = Str()
-        salary = Int()
-
-        traits_view = View(
-            Group(
-                Item(name='first_name'),
-                Item(name='last_name'),
-                Item(name='department'),
-                label='Personnel profile',
-                show_border=True,
-            ),
-        )
-
-        all_view = View(
-            Group(
-                Item(name='first_name'),
-                Item(name='last_name'),
-                Item(name='department'),
-                Item(name='employee_number'),
-                Item(name='salary'),
-                label='Personnel database entry',
-                show_border=True,
-            ),
-        )
-
-    sam = SimpleEmployee3()
-    sam.configure_traits()
-    sam.configure_traits(view='all_view')
+.. literalinclude:: examples/multiple_views.py
+   :start-at: multiple_views.py
 
 .. index:: traits_view attribute, configure_traits(); view parameter
